# Conflicts:
#	devdocs/nodejs-manual-spans.md
#	internal/test/integration/components/nodejsserver/app.js
#	pkg/internal/nodejs/spanbridge_test/scenario.js

+++ pkg/ebpf/common/go_channel_link_transform_test.go
@@ -19,8 +19,8 @@ import (
 )
 
 func TestGoChannelLinkEventTypeDoesNotConflictWithRuntimeMetrics(t *testing.T) {
-	assert.Equal(t, uint8(18), uint8(EventTypeGoChannelLink))
-	assert.NotEqual(t, uint8(17), uint8(EventTypeGoChannelLink))
+	assert.Equal(t, uint8(18), EventTypeGoChannelLink)
+	assert.NotEqual(t, uint8(17), EventTypeGoChannelLink)
 }
 
 func TestReadBPFTraceAsSpanGoChannelLinkEvent(t *testing.T) {
